+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Release v2.0
+
+* Major release - major changes
+* (Still) enforces combining third-party and app imports.
+* Module order ignores case (E.G. `str.lower()`)
+* Name order is case-sensitive (I.E. lexigraphical)
+* Relative module levels are reversed from default
+  * "import ..B" (level 2) is before "import .A" (level 1)
+* No longer deriving from Smarkets; using base Style class instead.
+* Added demo .flake8 and python file
+
 ## Release v1.1
 
 * Python 3.5 minimum.